feat: consolidate filename decryption logic and remove deprecated useDecryptFilenameOrg hook
